+++ jaden-backend/data/cart.json
@@ -1,9 +1,1 @@
-[
-  {
-    "id": "SyzsL_hJu",
-    "src": "/Images - Jaden/erys.jpg",
-    "name": "ERYS",
-    "price": 13.99,
-    "quantity": 7
-  }
-]
+[]
