+++ jaden-website/src/components/CartQuery.js
@@ -1,7 +1,56 @@
 import React, { useState, useRef } from 'react'
+import { useQuery, useMutation, useLazyQuery } from '@apollo/client'
+import { GET_CART } from '../graphql/Queries'
+import { PURCHASE_CART, REMOVE_FROM_CART, UPDATE_QUANTITY } from '../graphql/Mutations'
 import { Button } from './Button'
 
 function CartQuery(props) {
+
+    const { loading, error, data } = useQuery(GET_CART)
+    console.log(data)
+    if (data) { props.setCart(data)}
+
+    const [removeFromCart, 
+        { loading: mutationLoadingRemove, error: mutationErrorRemove }] = useMutation(REMOVE_FROM_CART, {
+        update: updateCartWithRemovedEntry
+    })
+
+    const [updateQuantity, 
+        { loading: mutationLoadingUpdate, error: mutationErrorUpdate }] = useMutation(UPDATE_QUANTITY, {
+            update: updateCartWithUpdatedEntry
+    })
+
+    function updateCartWithRemovedEntry(cache, { data }) {
+        cache.modify({
+            fields: {
+                allCart(existingCart = []) {
+                    const newCart = data.deleteCartItemById.cart
+                    console.log(newCart)
+                    console.log(...existingCart)
+                    cache.writeQuery({
+                        query: GET_CART,
+                        data: { newCart }
+                    })
+                }    
+            }
+        })
+    }
+
+    function updateCartWithUpdatedEntry(cache, { data }) {
+        cache.modify({
+            fields: {
+                allCart(existingCart = []) {
+                    const newCart = data.updateCartItemQuantityById.cart
+                    console.log(newCart)
+                    console.log(...existingCart)
+                    cache.writeQuery({
+                        query: GET_CART,
+                        data: { newCart }
+                    })
+                }    
+            }
+        })
+    }
 
     const [highlight, setHighlight] = useState()
     let refContainers = useRef([])
@@ -21,14 +70,10 @@ function CartQuery(props) {
         var parent = event.target.parentElement.parentElement.parentElement.getAttribute('data-key')
         console.log(parent)
         console.log(typeof(parent))
-        await props.removeFromCart({ variables: {
+        await removeFromCart({ variables: {
             idProvided: parent
         }})
     }
-
-    // function selectSize(size) {
-    //     console.log(size)
-    // }
 
     function changeQuantity(event) {
         var input = event.target
@@ -39,24 +84,18 @@ function CartQuery(props) {
         } else { 
             input.value = Math.round(input.value)
         }
-        props.updateQuantity({ variables: {
+        updateQuantity({ variables: {
             idProvided: id, newQuantity: parseInt(input.value)
         }})
     }
 
-    // function removeProductFromCart() {
-    //     console.log('')
-    // }
-
-    // buttonState="btn--select"
-
-    if (props.loading) return <h1 className="empty-cart">Loading...</h1>;
-    if (props.mutationLoadingRemove || props.mutationLoadingUpdate) return <h1 className="empty-cart">Removing From Cart...</h1>
-    if (props.error || props.mutationErrorRemove || props.mutationErrorUpdate) return <h1>Error! ${props.error}</h1>
-    if (props.data.allCart.length == 0) return <h1 className="empty-cart">Your cart is empty</h1>
+    if (loading) return <h1 className="empty-cart">Loading...</h1>;
+    if (mutationLoadingRemove || mutationLoadingUpdate) return <h1 className="empty-cart">Removing From Cart...</h1>
+    if (error || mutationErrorRemove || mutationErrorUpdate) return <h1>Error! ${props.error}</h1>
+    if (data.allCart.length == 0) return <h1 className="empty-cart">Your cart is empty</h1>
 
     return (
-        props.data.allCart.map((product, index) => (
+        data.allCart.map((product, index) => (
             <div className="cart-products"
             key={index} data-key={product.id}>
                 <div className="cart-product">
