+++ jaden-website/src/components/MerchQuery.js
@@ -1,13 +1,41 @@
 import React, { useRef, useMemo, useLayoutEffect } from 'react'
+import { useQuery, useMutation, useLazyQuery } from '@apollo/client'
+import { GET_MERCH, GET_CART } from '../graphql/Queries'
+import { ADD_TO_CART } from '../graphql/Mutations'
 import { Button } from './Button'
 
 function MerchQuery(props) {
+
+    const { loading, error, data, refetch, networkStatus } = useQuery(GET_MERCH)
+    console.log(data)
+
+    const [addToCart,
+        { loading: mutationLoadingAdd, error: mutationErrorAdd }] = useMutation(ADD_TO_CART, {
+        update: updateCartWithNewEntry
+    })
+
+    function updateCartWithNewEntry(cache, { data }) {
+        cache.modify({
+            fields: {
+                allCart(existingCart = []) {
+                    const newCart = data.addToCart.cart
+                    console.log(newCart)
+                    console.log(...existingCart)
+                    cache.writeQuery({
+                        query: GET_CART,
+                        data: { newCart }
+                    })
+                    console.log(cache.data)
+                }    
+            }
+        })
+    }
 
     async function checkId(event) {
         var parent = event.target.parentElement.parentElement.getAttribute('data-key')
         console.log(parent)
         console.log(typeof(parent))
-        await props.addToCart({ variables: {
+        await addToCart({ variables: {
             idProvided: parent
         }})
     }
@@ -39,14 +67,14 @@ function MerchQuery(props) {
     //     }
     // }, [scrollTo, props.data]);
 
-    if (props.networkStatus === 4) return <h1>Refetching</h1>
-    if (props.loading) return <h1>Loading...</h1>;
-    if (props.mutationLoading) return <h1 classname="empty-cart">Adding To Cart...</h1>
-    if (props.error || props.mutationError) return <h1>Error! ${props.error}</h1>
-    if (props.data.allMerch.length === 0) return <h1 className="empty-cart">Your Cart is Empty</h1>
+    if (networkStatus === 4) return <h1>Refetching</h1>
+    if (loading) return <h1>Loading...</h1>;
+    if (mutationLoadingAdd) return <h1 classname="empty-cart">Adding To Cart...</h1>
+    if (error || mutationErrorAdd) return <h1>Error! ${error}</h1>
+    if (data.allMerch.length === 0) return <h1 className="empty-cart">Your Cart is Empty</h1>
 
     return (
-        props.data.allMerch.map((merch, index) => (
+        data.allMerch.map((merch, index) => (
             <div className="merch-product" key={index} data-key={merch.id} data-item="true">
                 <img className="merch-image" alt="Merch" src={merch.src}></img>
                 <div className="merch-info">
